Move `impl Linked` to entity_linked module
