+++ src/tests_cfg/cake.rs
@@ -31,37 +31,4 @@ impl Related<super::filling::Entity> for Entity {
     }
 }
 
-#[derive(Debug)]
-pub struct CakeToFilling;
-
-impl Linked for CakeToFilling {
-    type FromEntity = Entity;
-
-    type ToEntity = super::filling::Entity;
-
-    fn link(&self) -> Vec<RelationDef> {
-        vec![
-            super::cake_filling::Relation::Cake.def().rev(),
-            super::cake_filling::Relation::Filling.def(),
-        ]
-    }
-}
-
-#[derive(Debug)]
-pub struct CakeToFillingVendor;
-
-impl Linked for CakeToFillingVendor {
-    type FromEntity = Entity;
-
-    type ToEntity = super::vendor::Entity;
-
-    fn link(&self) -> Vec<RelationDef> {
-        vec![
-            super::cake_filling::Relation::Cake.def().rev(),
-            super::cake_filling::Relation::Filling.def(),
-            super::filling::Relation::Vendor.def(),
-        ]
-    }
-}
-
 impl ActiveModelBehavior for ActiveModel {}
